Revised note on reactivity and `data`
The previous wording felt intimidating to me – if I didn't understand what that meant, I'd have to puzzle out what a reactivity system does. This change attempts to explain a little more up front – "starts watching for changes" instead of "adds properties to system" – and uses an example to show why we do it. It also links to Vue's reactivity deep dive for the student who wants to dig into it.